server/requirements.txt
fastapi
uvicorn
python-multipart
docling
mlx_vlm
PyMuPDF
Pillow>=9.0.0

